Autorun: Cleanup previous stored home path
The previous stored home path is unlikely the current home path of the
experiment if the experiment is loaded into Igor. The autorun feature
relyes on the correct home path to determine if an autorun should be
executed or not.

This has been fixed in such a way that the AfterFileOpenHook will always
clear the previous stored home path before checking the autorun
configuration.

Close: #457