treewide: simplify tracing support check
Simplify the Igor version check to check if tracing is supported. The simplified
code uses the compiler symbol `UTF_ALLOW_TRACING` to define and check if tracing
is supported or not.

The logic to define the compiler symbol `UTF_ALLOW_TRACING` takes more lines,
but it is easier to read and understand. The definition of such a symbol must be
done in each file, because Igor doesn't support the definition of such symbols
across files.