Tracing: Fix switch statement tracing to support IP10
With Igor 10 there is no code in specific regions of a switch statements allowed:

```igorpro
switch(condition)
	print "no code supported here"
	case VAR:
		print "normal code"
		break
		print "no code supported here"
endswitch
```

Our tracing logic added `Z_` function calls in such regions and therefore is a
logic added that prevents the addition of such. This fix does the following
changes to the coverage report:

- All comments, that are located in regions that do not support the execution of
  `Z_` functions (like certain regions in switch statements), are uncoverable.
- All compiler pragmas like `#if` and `#else` are now uncoverable.

Uncoverable lines are excluded from coverage report and will be rendered gray in
most visualizations.

This fix might change the overall code coverage slightly. This isn't the case if
you already had a code coverage of 100%.