Add WaveTracking for each test case and user hook
Before this change, wave tracking was done a single time for the test case and
end hook combined. The error was added to the end hook results after it was
published to the test suite results. This resulted in, that wave tracking errors
are ignored if we have a test case end hook.

Now is wave tracking done for each test case itself and for each user hook. The
error is always added to the correct entry and published to the test suite
results, which will then lead to a overall test fail.