Do wave leak checking only once for each test case
In the previous implementation was a wave leak check performed for each test
case and each user hook. Igortest does only check weak leaks once for a test
case in combination with its test case begin and end hook now. There is no
longer a separate weak leak check for a user hook, and test and test suite begin
and end hooks no longer have a weak leak check at all.

If a wave leak was found it will be reported as part of the test case user end
hook or if such does not exists to the test case itself.

The tests has been adapted to the changes and a new test case has been added
that check a scenario without any user hooks. The new test cases eliminate the
existing user end hook in its reentry function and remove the need of an
additional test file.